[Refactor](BE) Uniform Expr, Block, Columns size-relative interface (Part I) (#42930)
## Proposed changes
In this pr we refactored some base interface to avoid narrow cast with
data consistent kept.

We uniformed these type of numbers:
```
columns in Block        - U32 (just use no more than U16)
Rows in Block           - U64 (others U32, ColumnString64 U64)
column_id in expr arg   - I32 (because -1 is meaningful)
child of Expr           - U16 (FE use int32, have a restriction of 10000. adjustable but no guarantee for result)
```

So, for functions' argument column id, we use `uint32` to replace
`size_t`.
For rows, we use `size_t` to replace some `int`.
For child of expr, we cast it as `uint_16` to use, rather than `size_t`
as `vector::size()`

Refactor of table functions: https://github.com/apache/doris/pull/43218

+++ be/src/olap/column_predicate.h
@@ -325,7 +325,7 @@ protected:
         _judge_filter_rows = 0;
     }
 
-    void do_judge_selectivity(int64_t filter_rows, int64_t input_rows) const {
+    void do_judge_selectivity(uint64_t filter_rows, uint64_t input_rows) const {
         if ((_judge_counter--) == 0) {
             reset_judge_selectivity();
         }
@@ -352,8 +352,8 @@ protected:
     // without recalculating. At the beginning of the next period,
     // reset_judge_selectivity is used to reset these variables.
     mutable int _judge_counter = 0;
-    mutable int _judge_input_rows = 0;
-    mutable int _judge_filter_rows = 0;
+    mutable uint64_t _judge_input_rows = 0;
+    mutable uint64_t _judge_filter_rows = 0;
     mutable bool _always_true = false;
 };
